requirements.txt
openai
python-dotenv
flask
flask_cors
websockets
flask-sock
twilio
gunicorn
gevent
fastmcp
requests
httpx

celery[redis]